Add files via upload
ABSTRACT

Brain tumor is a life-threatening disease that requires early detection for effective treatment. In recent years, the use of deep learning algorithms has become a popular 
approach for brain tumor detection. In this paper, we propose a novel approach that combines Multi-Class CNN algorithms for brain tumor detection.Our approach uses 
magnetic resonance imaging (MRI) data as input and extracts features using the MultiClass CNN network. The extracted features are then fed into the Multi-Class CNN 
algorithm for classification. The proposed approach is evaluated on a publicly available brain tumor dataset, and the results show that our approach achieves high accuracy and outperforms other state-of-the-art methods.The Multi-Class CNN network is used to extract relevant temporal features from the MRI data, and the Multi-Class CNN 
algorithm is used for classification to improve the accuracy of the model. Multi-Class CNN is a machine learning algorithm that can be used to boost the performance of 
weak classifiers. The combination of these two algorithms results in a powerful model that can accurately detect brain tumors.In conclusion, our proposed approach provides 
a promising solution for brain tumor detection using deep learning algorithms. The combination of Multi-Class CNN algorithms offers a robust and accurate method for 
early detection of brain tumors, which can ultimately save lives.